+++ app/src/main/java/be/ugent/zeus/hydra/ui/main/resto/MenuLiveData.java
@@ -1,4 +1,4 @@
-package be.ugent.zeus.hydra.ui.resto.menu;
+package be.ugent.zeus.hydra.ui.main.resto;
 
 import android.content.Context;
 import android.content.SharedPreferences;
