+++ app/src/main/java/be/ugent/zeus/hydra/ui/main/resto/MenuViewModel.java
@@ -1,4 +1,4 @@
-package be.ugent.zeus.hydra.ui.resto.menu;
+package be.ugent.zeus.hydra.ui.main.resto;
 
 import android.app.Application;
