+++ app/src/main/res/layout/activity_main.xml
@@ -61,6 +61,16 @@
             <include android:visibility="gone" layout="@layout/x_progress_bar" android:id="@+id/drawer_loading" />
         </FrameLayout>
 
+        <!-- TODO: move this to fragment instead -->
+        <android.support.design.widget.BottomNavigationView
+            android:visibility="gone"
+            android:id="@+id/bottom_navigation"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_gravity="bottom"
+            app:menu="@menu/bottom_resto"
+            app:itemBackground="?android:colorBackground" />
+
     </android.support.design.widget.CoordinatorLayout>
 
     <!-- The navigation drawer that comes from the left -->
