--- app/src/main/res/layout/activity_resto.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.resto.menu.MenuActivity">
-
-    <android.support.design.widget.AppBarLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@id/app_bar_layout">
-
-        <android.support.v7.widget.Toolbar xmlns:android="http://schemas.android.com/apk/res/android"
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
-            android:ellipsize="marquee">
-
-            <Spinner
-                android:id="@+id/spinner"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
-            <ProgressBar
-                android:id="@+id/spinner_progress"
-                android:layout_width="24dp"
-                android:layout_height="24dp" />
-
-        </android.support.v7.widget.Toolbar>
-
-        <!-- TODO: investigate padding -->
-        <android.support.design.widget.TabLayout
-            android:id="@+id/resto_tabs_slider"
-            app:tabContentStart="@dimen/keyline_tab"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
-            app:tabMode="scrollable" />
-
-    </android.support.design.widget.AppBarLayout>
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_behavior="@string/appbar_scrolling_view_behavior">
-
-        <!-- The tabbed menu page -->
-        <android.support.v4.view.ViewPager
-            android:id="@+id/resto_tabs_content"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:animateLayoutChanges="true"
-            android:paddingTop="@dimen/content_spacing"
-            android:paddingLeft="@dimen/resto_menu_tab_spacing"
-            android:paddingRight="@dimen/resto_menu_tab_spacing" />
-
-        <!-- The progress bar while loading the menu -->
-        <include layout="@layout/x_progress_bar" />
-
-    </FrameLayout>
-
-</android.support.design.widget.CoordinatorLayout>
