+++ app/src/main/res/layout/fragment_resto.xml
@@ -1,108 +1,28 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<android.support.v4.widget.NestedScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:orientation="vertical">
 
-    <LinearLayout
+    <FrameLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:paddingLeft="@dimen/card_margin"
-        android:paddingRight="@dimen/card_margin"
-        android:paddingTop="@dimen/card_spacing">
+        android:layout_height="0dp"
+        android:layout_weight="1">
 
-        <android.support.v7.widget.CardView
-            android:animateLayoutChanges="true"
-            android:id="@+id/menu_today_card"
+        <!-- The tabbed menu page -->
+        <android.support.v4.view.ViewPager
+            android:id="@+id/resto_tabs_content"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:foreground="?android:attr/selectableItemBackground"
-            android:layout_marginBottom="@dimen/card_spacing"
-            android:clickable="true"
-            app:contentPaddingLeft="@dimen/card_title_padding_vertical"
-            app:contentPaddingRight="@dimen/card_title_padding_vertical">
+            android:layout_height="match_parent"
+            android:paddingTop="@dimen/content_spacing"
+            android:paddingLeft="@dimen/resto_menu_tab_spacing"
+            android:paddingRight="@dimen/resto_menu_tab_spacing" />
 
-            <!-- Display the menu of today -->
-            <LinearLayout
-                android:visibility="gone"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:orientation="vertical"
-                android:id="@+id/menu_today_card_layout"
-                android:animateLayoutChanges="true"
-                android:paddingBottom="@dimen/card_text_padding_bottom_no_content">
+        <!-- The progress bar while loading the menu -->
+        <include layout="@layout/x_progress_bar" />
 
-                <TextView
-                    android:id="@+id/menu_today_card_title"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:paddingTop="@dimen/card_title_padding_top_large_title"
-                    android:paddingBottom="@dimen/card_title_padding_bottom"
-                    style="@style/Base.TextAppearance.AppCompat.Title"
-                    android:textSize="@dimen/card_title_size_large"
-                    tools:text="Menutitel" />
+    </FrameLayout>
 
-                <be.ugent.zeus.hydra.ui.common.widgets.MenuTable
-                    android:id="@+id/menu_table"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    app:show="all" />
-
-            </LinearLayout>
-
-            <include layout="@layout/x_progress_bar" />
-
-            <TextView
-                android:visibility="gone"
-                android:paddingTop="@dimen/card_text_padding_top"
-                android:paddingBottom="@dimen/card_text_padding_bottom_no_content"
-                android:id="@+id/error_view"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/resto_no_menu" />
-
-        </android.support.v7.widget.CardView>
-
-        <be.ugent.zeus.hydra.ui.common.widgets.ButtonBarLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center_horizontal">
-
-            <!-- We cannot set the drawables here, since before Lollipop, there is no support for
-            vectors. The support library used to support this, but removed it since it caused
-            memory problems. -->
-
-            <Button
-                android:id="@+id/home_resto_view"
-                android:layout_height="wrap_content"
-                android:layout_width="wrap_content"
-                android:text="@string/home_resto_view_menu"
-                tools:drawableTop="@drawable/btn_restaurant_menu"
-                android:drawableTint="?android:colorControlNormal"
-                android:maxLines="1" />
-
-            <Button
-                android:id="@+id/home_resto_view_sandwich"
-                android:layout_height="wrap_content"
-                android:layout_width="wrap_content"
-                android:text="@string/home_resto_view_sandwiches"
-                tools:drawableTop="@drawable/btn_sandwich"
-                android:drawableTint="?android:colorControlNormal"
-                android:maxLines="1" />
-
-            <Button
-                android:id="@+id/home_resto_view_resto"
-                android:layout_height="wrap_content"
-                android:layout_width="wrap_content"
-                android:text="@string/home_resto_view_resto"
-                tools:drawableTop="@drawable/btn_explore"
-                android:drawableTint="?android:colorControlNormal"
-                android:maxLines="1" />
-
-        </be.ugent.zeus.hydra.ui.common.widgets.ButtonBarLayout>
-
-    </LinearLayout>
-</android.support.v4.widget.NestedScrollView>
+</LinearLayout>
