--- app/src/main/res/layout/fragment_resto2.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <!-- The tabbed menu page -->
-    <android.support.v4.view.ViewPager
-        android:id="@+id/resto_tabs_content"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:paddingTop="@dimen/content_spacing"
-        android:paddingLeft="@dimen/resto_menu_tab_spacing"
-        android:paddingRight="@dimen/resto_menu_tab_spacing" />
-
-    <!-- The progress bar while loading the menu -->
-    <include layout="@layout/x_progress_bar" />
-
-</FrameLayout>
