+++ app/src/main/res/menu/bottom_resto.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/resto_bottom_sandwich"
+        android:icon="@drawable/btn_sandwich"
+        android:title="@string/home_resto_view_sandwiches"
+        app:showAsAction="ifRoom" />
+
+    <item
+        android:id="@+id/resto_bottom_locations"
+        android:icon="@drawable/btn_explore"
+        android:title="@string/home_resto_view_resto"
+        app:showAsAction="ifRoom" />
+
+</menu>
